Changing objects queries location
